+++ src/component/Anchor/Anchor.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link, LinkProps } from "@remix-run/react";
 import { Anchor as MantineAnchor, AnchorProps as MantineAnchorProps } from "@mantine/core";
-import { Optional } from "@encode42/mantine-extras";
+import { Optional } from "@encode42/node-extras";
 
 export type AnchorProps = MantineAnchorProps & Optional<LinkProps, "to">;