+++ src/util/interface/api/APIRoute.ts
@@ -1,3 +1,6 @@
+/**
+ * Ease-of-use interface for adding the `apiRoute` property to a component.
+ */
 export interface APIRoute {
-    "apiRoute": string
+    "apiRoute"?: string
 }